+++ client/Packages/com.beamable/Common/Runtime/Content/Optionals.cs
@@ -3,6 +3,7 @@ using System;
 using System.Collections;
 using System.Collections.Generic;
 using System.ComponentModel;
+using System.Diagnostics;
 using System.Globalization;
 #pragma warning disable CS0618
 
@@ -39,6 +40,8 @@ namespace Beamable.Common.Content
 
 	[System.Serializable]
 	[Agnostic]
+	
+	[DebuggerDisplay("{HasValue ? (Value?.ToString()) : \"no value\"}")]
 	public class Optional<T> : Optional
 	{
 		public static implicit operator T(Optional<T> option) => option?.HasValue == true ? (T)option.Value : default(T);
@@ -89,13 +92,29 @@ namespace Beamable.Common.Content
 			if (!HasValue) throw exFactory?.Invoke() ?? new ArgumentException("Optional value does not exist, but it was forced.");
 			return Value;
 		}
-
+		
 		public T GetOrElse(T otherwise) => GetOrElse(() => otherwise);
 
 		public T GetOrElse(Func<T> otherwise)
 		{
 			if (HasValue) return Value;
 			return otherwise();
+		}
+
+		public TResultOptional Map<TResult, TResultOptional>(Func<T, TResult> func, bool allowNull = false)
+			where TResultOptional : Optional<TResult>, new()
+			// where TResult
+		{
+			var result = new TResultOptional();
+			if (HasValue)
+			{
+				var resultVal = func(Value);
+				if (allowNull || resultVal != null)
+				{
+					result.Set(resultVal);
+				}
+			}
+			return result;
 		}
 
 		public Optional<T> DoIfExists(Action<T> callback)
@@ -135,6 +154,7 @@ namespace Beamable.Common.Content
 
 	[System.Serializable]
 	[Agnostic]
+	[DebuggerDisplay("{HasValue ? Value.ToString() : \"no value\"}")]
 	public class OptionalValue<T> : Optional<T> where T : struct
 	{
 		public static implicit operator T?(OptionalValue<T> option) => option?.HasValue == true ? (T?)option.Value : null;
@@ -266,6 +286,15 @@ namespace Beamable.Common.Content
 			Value = value;
 			HasValue = true;
 		}
+		
+		public string GetNonEmptyOrElse(Func<string> otherwise)
+		{
+			if (HasNonEmptyValue) return Value;
+			return otherwise();
+		}
+
+
+		public bool HasNonEmptyValue => HasValue && !string.IsNullOrEmpty(Value);
 	}
 
 	[System.Serializable]