+++ client/Packages/com.beamable/Editor/ResetPlayerAccessTokenEditor.cs
@@ -1,4 +1,5 @@
 using Beamable.Api;
+using System.Collections.Generic;
 using UnityEditor;
 using UnityEngine;
 using static Beamable.Common.Constants.MenuItems.Windows;
@@ -27,6 +28,9 @@ namespace Beamable.Editor
 				}
 				if (GUILayout.Button("Clear Token"))
 				{
+					BeamEditorContext.Default.EditorAccountService.editorAccounts = new List<EditorAccountInfo>();
+					BeamEditorContext.Default.EditorAccountService.cid.Clear();
+					BeamEditorContext.Default.Requester.DeleteToken();
 					var api = BeamEditorContext.Default;
 					var storage = new AccessTokenStorage(Prefix);
 					storage.ClearDeviceRefreshTokens(api.CurrentCustomer.Cid, api.CurrentRealm.Pid);